push

+++ TWP.Api.Core/DataTransferObjects/Monster5eDto.cs
@@ -6,8 +6,8 @@
         public string Name { get; set; } = string.Empty;
         public string? Alignment { get; set; }
         public string ChallengeRating { get; set; } = string.Empty;
-        public int Cr { get; set; }
-        public int? CrInLair { get; set; }
+        public float Cr { get; set; }
+        public float? CrInLair { get; set; }
         public int Xp { get; set; }
         public int InitiativeBonus { get; set; }
         public string? Role { get; set; }
